Crate primary

Re-exports

pub use crate::block_synchronizer::mock::MockBlockSynchronizer;

pub use crate::block_synchronizer::responses::CertificatesResponse;

pub use crate::block_synchronizer::responses::PayloadAvailabilityResponse;

pub use crate::block_synchronizer::BlockHeader;

Modules

block_synchronizer

Structs

BlockRemover

BlockRemover is responsible for removing blocks identified by their certificate id (digest) from across our system. On high level It will make sure that the DAG is updated, internal storage where there certificates and headers are stored, and the corresponding batches as well.

BlockWaiter

BlockWaiter is responsible for fetching the block data from the downstream worker nodes. A block is basically the aggregate of batches of transactions for a given certificate.

DeleteBatchMessage

EndpointMetrics

GetBlockResponse

Primary

PrimaryChannelMetrics

Enums

BlockCommand

BlockRemoverCommand

NetworkModel

The network model in which the primary operates.

PrimaryWorkerMessage

The messages sent by the primary to its workers.

Constants

CHANNEL_CAPACITY

The default channel capacity for each channel of the primary.

Type Definitions

PayloadToken