Module primary::block_synchronizer::handler

Structs

BlockSynchronizerHandler

A helper struct to allow us access the block_synchronizer in a synchronous way. It also offers methods to both fetch the certificates and way to process them and causally complete their history.

Enums

Error

The errors returned by the Handler. It translates also the errors returned from the block_synchronizer.

Traits

Handler

Handler defines an interface to allow us access the BlockSycnhronizer’s functionality in a synchronous way without having to deal with message emission. The BlockSynchronizer on its own for the certificates that fetches on the fly from peers doesn’t care/deal about any other validation checks than the basic verification offered via the certificate entity it self. For that reason the Handler offers methods to submit the fetched from peers certificates for further validation & processing (e.x ensure parents history is causally complete) to the core and wait until the certificate has been processed, before it returns it back as result.