Struct primary::PrimaryChannelMetrics

pub struct PrimaryChannelMetrics {

    pub tx_others_digests: IntGauge,
    pub tx_our_digests: IntGauge,
    pub tx_parents: IntGauge,
    pub tx_headers: IntGauge,
    pub tx_sync_headers: IntGauge,
    pub tx_sync_certificates: IntGauge,
    pub tx_headers_loopback: IntGauge,
    pub tx_certificates_loopback: IntGauge,
    pub tx_primary_messages: IntGauge,
    pub tx_helper_requests: IntGauge,
    pub tx_get_block_commands: IntGauge,
    pub tx_batches: IntGauge,
    pub tx_block_removal_commands: IntGauge,
    pub tx_batch_removal: IntGauge,
    pub tx_block_synchronizer_commands: IntGauge,
    pub tx_certificate_responses: IntGauge,
    pub tx_payload_availability_responses: IntGauge,
    pub tx_state_handler: IntGauge,
    pub tx_reconfigure: IntGauge,
    pub tx_committed_certificates: IntGauge,
    pub tx_new_certificates: IntGauge,

}

Fields
tx_others_digests: IntGauge

occupancy of the channel from the primary::WorkerReceiverHandler to the primary::PayloadReceiver

tx_our_digests: IntGauge

occupancy of the channel from the primary::WorkerReceiverHandler to the primary::Proposer

tx_parents: IntGauge

occupancy of the channel from the primary::Core to the primary::Proposer

tx_headers: IntGauge

occupancy of the channel from the primary::Proposer to the primary::Core

tx_sync_headers: IntGauge

occupancy of the channel from the primary::Synchronizer to the primary::HeaderWaiter

tx_sync_certificates: IntGauge

occupancy of the channel from the primary::Synchronizer to the primary::CertificaterWaiter

tx_headers_loopback: IntGauge

occupancy of the channel from the primary::HeaderWaiter to the primary::Core

tx_certificates_loopback: IntGauge

occupancy of the channel from the primary::CertificateWaiter to the primary::Core

tx_primary_messages: IntGauge

occupancy of the channel from the primary::PrimaryReceiverHandler to the primary::Core

tx_helper_requests: IntGauge

occupancy of the channel from the primary::PrimaryReceiverHandler to the primary::Helper

tx_get_block_commands: IntGauge

occupancy of the channel from the primary::ConsensusAPIGrpc (when external consensus is being used) & executor::Subscriber (when internal consensus, ex Bullshark, is being used) to the primary::BlockWaiter.

tx_batches: IntGauge

occupancy of the channel from the primary::WorkerReceiverHandler to the primary::BlockWaiter

tx_block_removal_commands: IntGauge

occupancy of the channel from the primary::ConsensusAPIGrpc to the primary::BlockRemover

tx_batch_removal: IntGauge

occupancy of the channel from the primary::WorkerReceiverHandler to the primary::BlockRemover

tx_block_synchronizer_commands: IntGauge

occupancy of the channel from the primary::BlockSynchronizerHandler to the primary::BlockSynchronizer

tx_certificate_responses: IntGauge

occupancy of the channel from the primary::PrimaryReceiverHandler to the primary::BlockSynchronizer

tx_payload_availability_responses: IntGauge

occupancy of the channel from the primary::PrimaryReceiverHandler to the primary::BlockSynchronizer

tx_state_handler: IntGauge

occupancy of the channel from the primary::WorkerReceiverHandler to the primary::StateHandler

tx_reconfigure: IntGauge

occupancy of the channel from the reconfigure notification to most components.

tx_committed_certificates: IntGauge

occupancy of the channel from the Consensus to the primary::Core

tx_new_certificates: IntGauge

occupancy of the channel from the primary::Core to the Consensus

Implementations

impl PrimaryChannelMetrics

pub const NAME_COMMITTED_CERTS: &'static str = "tx_committed_certificates"

pub const DESC_COMMITTED_CERTS: &'static str = "occupancy of the channel from the `Consensus` to the `primary::Core`"

pub const NAME_NEW_CERTS: &'static str = "tx_new_certificates"

pub const DESC_NEW_CERTS: &'static str = "occupancy of the channel from the `primary::Core` to the `Consensus`"

pub const NAME_GET_BLOCK_COMMANDS: &'static str = "tx_get_block_commands"

pub const DESC_GET_BLOCK_COMMANDS: &'static str = "occupancy of the channel from the `primary::ConsensusAPIGrpc` & `executor::Subscriber` to the `primary::BlockWaiter`"

pub fn new(registry: &Registry) -> Self

pub fn replace_registered_new_certificates_metric(
&mut self,
registry: &Registry,
collector: Box<GenericGauge<AtomicI64>>
)

pub fn replace_registered_committed_certificates_metric(
&mut self,
registry: &Registry,
collector: Box<GenericGauge<AtomicI64>>
)

pub fn replace_registered_get_block_commands_metric(
&mut self,
registry: &Registry,
collector: Box<GenericGauge<AtomicI64>>
)